refactor: split out other plugins from vite.config.js
