+++ vite.config.js
@@ -1,9 +1,8 @@
-import { spawnSync } from "node:child_process";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { defineConfig } from "vite";
-import { DynamicPublicDirectory } from "vite-multiple-assets";
-import { run } from "vite-plugin-run";
+import { getSpritesheetsPlugin } from "./vite/get-spritesheets-plugin.js";
+import { vitePluginBundledCssAfterBulma } from "./vite/vite-plugin-bundled-css-after-bulma.js";
 import {
   itemMetadataPlugins,
   itemMetadataResolveAliases,
@@ -12,143 +11,10 @@ import {
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 /**
- * Item-metadata pipeline (plan step 4): `vite/wiring.mjs` registers the pre-plugin and
+ * Item-metadata pipeline (plan step 4): `vite/wiring.js` registers the pre-plugin and
  * `resolve.alias` for both **serve** and **build**. Testem’s Vite middleware loads this file
  * via `configFile` so browser tests get the same behavior. Other plugins stay below.
  */
-
-/** @returns {string[]} Command and args for vite-plugin-run (first element is the executable). */
-function copySpritesheetsRsyncRun() {
-  return [
-    "rsync",
-    "-ahu",
-    "--delete",
-    "--info=progress2",
-    "--no-inc-recursive",
-    "spritesheets",
-    "dist",
-  ];
-}
-
-/**
- * Windows: mirror `spritesheets/` into `dist/spritesheets` with robocopy (same flags as before).
- * Robocopy uses exit codes 0–7 for success; ≥8 is failure — we fail the build only on real errors.
- */
-function vitePluginCopySpritesheetsRobocopy() {
-  return {
-    name: "copy-spritesheets-robocopy",
-    apply: "build",
-    closeBundle() {
-      const dest = path.join("dist", "spritesheets");
-      const result = spawnSync(
-        "robocopy",
-        [
-          "spritesheets",
-          dest,
-          "/MIR",
-          "/Z",
-          "/XO",
-          "/MT:8",
-          "/NFL",
-          "/NDL",
-          "/NJH",
-          "/NJS",
-          "/NP",
-        ],
-        { stdio: "inherit", shell: false, windowsHide: true },
-      );
-      if (result.error) {
-        throw result.error;
-      }
-      const code = result.status;
-      if (code === null) {
-        throw new Error("robocopy was terminated by a signal");
-      }
-      if (code >= 8) {
-        throw new Error(`robocopy failed with exit code ${code}`);
-      }
-    },
-  };
-}
-
-/**
- * Plugin that keeps spritesheets available to Vite: in dev, serve the tree from disk; on build,
- * copy (mirror) `spritesheets/` into `dist/spritesheets` so the production output matches the repo.
- *
- * - **Dev (`vite`, `command === "serve"`):** serve `public/` and `spritesheets/` (no `dist/` copy).
- * - **Build on Windows:** robocopy into `dist/` with exit codes mapped so real failures fail the build.
- * - **Build on macOS / Linux:** `rsync` via `vite-plugin-run`.
- *
- * @param {"serve" | "build"} command Vite CLI command from `defineConfig`.
- */
-function getSpritesheetsPlugin(command) {
-  if (command === "serve") {
-    return DynamicPublicDirectory(["public/**", "{\x01,spritesheets}/**"]);
-  }
-
-  if (process.platform === "win32") {
-    return vitePluginCopySpritesheetsRobocopy();
-  }
-
-  return run({
-    input: [
-      {
-        name: "copy spritesheets",
-        run: copySpritesheetsRsyncRun,
-        condition: () => true,
-        onFileChanged: () => {},
-      },
-    ],
-    silent: false,
-  });
-}
-
-/**
- * Vite injects the extracted entry CSS after module scripts in built HTML.
- * Keep Bulma first, then the app bundle (same cascade as source index.html).
- */
-function vitePluginBundledCssAfterBulma() {
-  return {
-    name: "bundled-css-after-bulma",
-    transformIndexHtml: {
-      order: "post",
-      handler(html) {
-        const stylesheetLinkRe = /<link\b[^>]*\brel=["']stylesheet["'][^>]*>/gi;
-        const bundled = [];
-        let m;
-        while ((m = stylesheetLinkRe.exec(html)) !== null) {
-          const tag = m[0];
-          // With `base: "./"`, href is `./assets/...`; older builds used `/assets/...`.
-          if (
-            /assets\/[^"'>\s]+\.css\b/.test(tag) &&
-            !/https?:\/\//i.test(tag)
-          ) {
-            bundled.push(tag);
-          }
-        }
-        if (bundled.length === 0) {
-          return html;
-        }
-
-        let out = html;
-        for (const tag of bundled) {
-          out = out.replace(tag, "");
-        }
-
-        const bulmaRe =
-          /(<link\b[^>]*\brel=["']stylesheet["'][^>]*bulma[^>]*>)/i;
-        if (!bulmaRe.test(out)) {
-          return html;
-        }
-
-        return out.replace(
-          bulmaRe,
-          (_, bulmaTag) => `${bulmaTag}\n\t${bundled.join("\n\t")}`,
-        );
-      },
-    },
-  };
-}
 
 export default defineConfig(({ command }) => ({
   base: "./",
